requirements.txt
asgiref==3.7.2
Django==3.2
django-admin-thumbnails==0.2.8
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.12.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
